REFACTOR: Divide component
